KYAN-177 extract AuthorInfoModel
