+++ applications/blogs/backend/src/main/resources/libs/kyanite/blogs/components/blogarticlepage/dialog/.content.json
@@ -3,33 +3,9 @@
     "metadata": {
       "sling:resourceType": "wcm/dialogs/components/tab",
       "label": "Metadata",
-      "authorName": {
-        "sling:resourceType": "wcm/dialogs/components/textfield",
-        "name": "authorName",
-        "label": "Author name"
-      },
-      "authorPhoto": {
-        "sling:resourceType": "wcm/dialogs/components/pathpicker",
-        "rootPath": "/content",
-        "name": "authorPhoto",
-        "description": "Author photo",
-        "label": "Author photo"
-      },
-      "authorPhotoAlt": {
-        "sling:resourceType": "wcm/dialogs/components/textfield",
-        "name": "authorPhotoAlt",
-        "description": "Author photo alt",
-        "label": "Author photo alt"
-      },
-      "authorRole": {
-        "sling:resourceType": "wcm/dialogs/components/textfield",
-        "name": "authorRole",
-        "label": "Author role"
-      },
-      "authorDescription": {
-        "sling:resourceType": "wcm/dialogs/components/textfield",
-        "name": "authorDescription",
-        "label": "Author description"
+      "author": {
+        "sling:resourceType": "wcm/dialogs/components/include",
+        "path": "/libs/kyanite/blogs/components/common/dialog/authorinfo"
       },
       "publicationDate": {
         "sling:resourceType": "wcm/dialogs/components/datepicker",
